refactor: `Work` に関する操作をより適当な場所に配置
