+++ src/context/work/types.ts
@@ -18,12 +18,3 @@ export type WorkType = (typeof WorkType)[number];
 
 export const WorkContext = ['commission', 'collaboration', 'corporate', 'personal'] as const;
 export type WorkContext = (typeof WorkContext)[number];
-
-export const getTopThumbnailUrl = (work: Work): string => {
-	const topThumbnail = work.visualImageUrl.at(0);
-	if (topThumbnail == null) {
-		throw new Error(`Expected to have at least one thumnbail for work '${work.id}'`);
-	}
-
-	return topThumbnail;
-};
